delete case-quizblock and re-introduce it as a simple subclass with a few method overrides
