+++ uelc/main/models.py
@@ -1,9 +1,11 @@
 from django.db import models
 from django import forms
 from django.contrib.auth.models import User
-from pagetree.models import Hierarchy, Section
+from pagetree.models import Hierarchy, Section, ReportableInterface
 from pageblocks.models import TextBlock
-
+from quizblock.models import Quiz, Submission, Response
+from gate_block.models import GateSubmission
+from django.core.exceptions import ObjectDoesNotExist
 
 class Cohort(models.Model):
     name = models.CharField(max_length=255, blank=False)
@@ -280,3 +282,115 @@ class LibraryItem(models.Model):
                            }))
 
         return EditLibraryForm()
+
+
+class CaseQuiz(Quiz):
+    display_name = "Case Quiz"
+    template_file = "quizblock/quizblock.html"
+
+    @classmethod
+    def get_pageblock(self):
+        return True
+
+    @classmethod
+    def create(self, request):
+        return CaseQuiz.objects.create(
+            description=request.POST.get('description', ''),
+            rhetorical=request.POST.get('rhetorical', ''),
+            allow_redo=request.POST.get('allow_redo', ''),
+            show_submit_state=request.POST.get('show_submit_state', False))
+
+    @classmethod
+    def create_from_dict(self, d):
+        q = CaseQuiz.objects.create(
+            description=d.get('description', ''),
+            rhetorical=d.get('rhetorical', False),
+            allow_redo=d.get('allow_redo', True),
+            show_submit_state=d.get('show_submit_state', True)
+        )
+        q.import_from_dict(d)
+        return q
+
+    @classmethod
+    def add_form(self):
+        class AddForm(forms.Form):
+            description = forms.CharField(widget=forms.widgets.Textarea())
+            rhetorical = forms.BooleanField()
+            allow_redo = forms.BooleanField()
+            show_submit_state = forms.BooleanField(initial=True)
+        return AddForm()
+
+    def submit(self, user, data):
+        """ a big open question here is whether we should
+        be validating submitted answers here, on submission,
+        or let them submit whatever garbage they want and only
+        worry about it when we show the admins the results """
+        s = Submission.objects.create(quiz=self, user=user)
+        # create a CaseMap for the user
+        # get Case the user is currently on
+        quiz = self
+        for k in data.keys():
+            if k.startswith('case'):
+                case_id = data[k]
+                try:
+                    casemap = CaseMap.objects.get(
+                        user=user,
+                        case_id=case_id)
+
+                except ObjectDoesNotExist:
+                    casemap = CaseMap.objects.create(
+                        user=user,
+                        case_id=case_id,
+                        value=str(0))
+
+                casemap.set_value(quiz, data)
+
+            if k.startswith('question'):
+                qid = int(k[len('question'):])
+                question = CaseQuestion.objects.get(id=qid)
+                # it might make more sense to just accept a QueryDict
+                # instead of a dict so we can use getlist()
+                if isinstance(data[k], list):
+                    for v in data[k]:
+                        Response.objects.create(
+                            submission=s,
+                            question=question,
+                            value=v)
+                else:
+                    Response.objects.create(
+                        submission=s,
+                        question=question,
+                        value=data[k])
+
+    def is_submitted(self, quiz, user):
+        return Submission.objects.filter(
+            quiz=self,
+            user=user).count() > 0
+
+    def unlocked(self, user, section):
+        # meaning that the user can proceed *past* this one,
+        # not that they can access this one. careful.
+        unlocked = False
+        submissions = GateSubmission.objects.filter(gate_user_id=user.id)
+        if len(submissions) > 0:
+            for sub in submissions:
+                section_id = sub.gateblock.pageblock().section_id
+                section = Section.objects.get(id=section_id)
+                upv = section.get_uservisit(user)
+                blocks = section.pageblock_set.all()
+                for block in blocks:
+                    obj = block.content_object
+                    if obj.display_name == "Gate Block":
+                        unlocked = obj.unlocked(user, section)
+            is_quiz_submitted = self.is_submitted(self, user)
+            if not (unlocked and is_quiz_submitted):
+                unlocked = False
+                upv.status = 'complete'
+                upv.save()
+            else:
+                upv.status = 'complete'
+                upv.save()
+                unlocked = True
+        return unlocked
+
+ReportableInterface.register(CaseQuiz)
